+++ src/client/results/Results.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import MapGL from 'react-map-gl';
-var ScatterPlotOverlay = require('react-map-gl/src/overlays/scatterplot.react');
-//require('script!mapbox-gl/dist/mapbox-gl-dev.js');
 import Immutable from 'immutable';
 import Feed from './Feed';
 import MostUsedWords from './mostfrequent/words/MostUsedWords';
 import MostActiveUsers from './mostfrequent/users/MostActiveUsers';
+import GoogleMap from './GoogleMap';
 
 let Results = ({ feed, mostFrequent }) => {
   const posts = feed.posts;
@@ -48,7 +46,7 @@ let Results = ({ feed, mostFrequent }) => {
         />
 
         <div id="tweetMap" style={{height: '500px', width:'100%'}} />
-        <GMap posts={posts.filter((post) => post.data.longitude !== 0)} />
+        <GoogleMap posts={posts.filter((post) => post.data.longitude !== 0)} />
 
       </div>
 
@@ -62,85 +60,6 @@ let Results = ({ feed, mostFrequent }) => {
 
   );
 };
-
-
-const GMap = React.createClass({
-  // http://revelry.co/google-maps-react/
-  map: null,
-  markers: [],
-  infoWindow: null,
-
-  render: () => (
-    <div />
-  ),
-
-  componentDidUpdate: function() {
-    console.log("update");
-    this.markerBounds = new google.maps.LatLngBounds();
-    this.markers.forEach((marker) => marker.setMap(null));
-    this.props.posts.forEach(this.addMarker);
-    this.map.fitBounds(this.markerBounds);
-  },
-
-  componentDidMount: function() {
-    this.map = this.createMap()
-    this.componentDidUpdate();
-  },
-
-  createMap: function() {
-    const mapOptions = {
-      minZoom: 1,
-      zoom: 1,
-      center: new google.maps.LatLng(30, 0),
-      mapTypeId: google.maps.MapTypeId.TERRAIN
-    }
-
-    console.log("map", mapOptions);
-    //return new google.maps.Map(this.refs.map_canvas.getDOMNode(), mapOptions)
-    return new google.maps.Map($("#tweetMap")[0], mapOptions)
-  },
-
-  addMarker: function(post) {
-    console.log("marker", post);
-    const pos = new google.maps.LatLng(post.data.latitude, post.data.longitude);
-    const marker = new google.maps.Marker({
-      position: pos,
-      map: this.map,
-      title: post.data.content
-    })
-
-    var infowindow = new google.maps.InfoWindow({
-      content: post.data.content
-    });
-
-    marker.addListener('click', function() {
-      infowindow.open(this.map, marker);
-    })
-
-    this.markers.push(marker);
-    this.markerBounds.extend(pos)
-
-    return marker;
-  },
-
-  createInfoWindow: function(marker) {
-    console.log("info");
-    const contentString = "<div class='InfoWindow'>I'm a Window that contains Info Yay</div>";
-    const infoWindow = new google.maps.InfoWindow({
-      map: this.map,
-      anchor: marker,
-      content: contentString
-    })
-    return infoWindow
-  },
-
-  handleZoomChange: () => {
-  },
-
-  handleDragEnd: () => {
-  },
-})
-
 
 const mapStateToProps = (state) => ({
   feed: state.feed,
